Move translator list utility to release-tool
